refactor: queries 폴더의 search, bookmark를 notice로 파일 통합
